sperate files, fix sigsegv bug, and add method support
